backend/requirements.txt
python-dotenv

uvicorn>=0.30.6
uvicorn-worker>=0.2.0
gunicorn>=23.0.0

Django>=5.1.1
djangorestframework>=3.15.2
django-cors-headers>=4.4.0
djangorestframework-simplejwt>=5.3.1
google-api-python-client>=2.146.0

django-redis>=5.4.0
redis>=5.0.8
psycopg>=3.2.2
psycopg-binary>=3.2.2

ipdb>=0.13.13
